DI-765 Add getVersion method to IRCTApplication class
A new method to return the version that was used during `mvn install`
goal. The build version, from the pom.xml, will be placed in a resource
file, during packaging. This file contains a value `X.X` for this
field, during testing. The supplied test will use this value to pass,
during unit testing.